// dashboard/src/client/models/node-result-delta-event.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Scheduler Public API (v1)
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.3.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { JsonPatchOperation } from './json-patch-operation';

export interface NodeResultDeltaEvent {
    'kind': NodeResultDeltaEventKindEnum;
    'runId': string;
    'nodeId': string;
    'revision': number;
    'sequence': number;
    'operation': NodeResultDeltaEventOperationEnum;
    'path'?: string | null;
    'payload'?: { [key: string]: any; } | null;
    'patches'?: Array<JsonPatchOperation> | null;
    'chunkMeta'?: { [key: string]: any; } | null;
    'terminal'?: boolean;
}

export const NodeResultDeltaEventKindEnum = {
    NodeResultDelta: 'node.result.delta'
} as const;

export type NodeResultDeltaEventKindEnum = typeof NodeResultDeltaEventKindEnum[keyof typeof NodeResultDeltaEventKindEnum];
export const NodeResultDeltaEventOperationEnum = {
    Append: 'append',
    Replace: 'replace',
    Patch: 'patch'
} as const;

export type NodeResultDeltaEventOperationEnum = typeof NodeResultDeltaEventOperationEnum[keyof typeof NodeResultDeltaEventOperationEnum];