// dashboard/src/client/models/node-result-snapshot-event.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Scheduler Public API (v1)
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.3.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { RunArtifact } from './run-artifact';

export interface NodeResultSnapshotEvent {
    'kind': NodeResultSnapshotEventKindEnum;
    'runId': string;
    'nodeId': string;
    'revision': number;
    'format'?: NodeResultSnapshotEventFormatEnum;
    'content': { [key: string]: any; };
    'artifacts'?: Array<RunArtifact> | null;
    'summary'?: string | null;
    'complete'?: boolean;
}

export const NodeResultSnapshotEventKindEnum = {
    NodeResultSnapshot: 'node.result.snapshot'
} as const;

export type NodeResultSnapshotEventKindEnum = typeof NodeResultSnapshotEventKindEnum[keyof typeof NodeResultSnapshotEventKindEnum];
export const NodeResultSnapshotEventFormatEnum = {
    Json: 'json',
    Binary: 'binary',
    Text: 'text'
} as const;

export type NodeResultSnapshotEventFormatEnum = typeof NodeResultSnapshotEventFormatEnum[keyof typeof NodeResultSnapshotEventFormatEnum];